chore: a try to bring things together
