+++ dash-abci/src/abci/messages.rs
@@ -33,7 +33,7 @@ pub struct BlockEndRequest {
 
 pub type EpochRefund = (u16, u64);
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct FeesAggregate {
     pub processing_fees: u64,
